I divided code to separate files.
I divided code to separate files and fixed one bug in
Sapper::init_game()
