+++ Translator.cpp
@@ -1,19 +1,4 @@
-#ifndef TRANSLATIONS_H
-#define TRANSLATIONS_H
-
-class Translator
-{
-private:
-    std::string language;
-    std::string PL(std::string message);
-    std::string EN(std::string message);
-public:
-    std::string get_language();
-    void set_language(std::string lang);
-    std::string loc(std::string message);
-    std::string to_en(std::string message);
-    std::string translate(std::string message);
-};
+#include "Translator.h"
 
 std::string Translator::get_language()
 {
@@ -97,5 +82,3 @@ std::string Translator::EN(std::string message)
 {
     return message;
 }
-
-#endif // TRANSLATIONS
